rag_chatbot/requirements.txt
# RAG Enterprise Chatbot - Python Dependencies
# Production-ready with vLLM integration

# ===========================================
# Web Framework
# ===========================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
starlette==0.35.1

# ===========================================
# Document Processing
# ===========================================
pypdf==3.17.4
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23

# ===========================================
# Text Processing
# ===========================================
tiktoken==0.5.2

# ===========================================
# Embeddings & ML
# ===========================================
sentence-transformers==2.2.2
torch>=2.0.0
transformers>=4.36.0

# ===========================================
# Vector Databases
# ===========================================
chromadb==0.4.22
qdrant-client==1.7.0

# ===========================================
# LLM Client (OpenAI-compatible for vLLM)
# ===========================================
openai==1.10.0
aiohttp>=3.9.0

# ===========================================
# Caching
# ===========================================
redis==5.0.1
aioredis==2.0.1

# ===========================================
# Resilience & Rate Limiting
# ===========================================
tenacity==8.2.3
slowapi==0.1.9

# ===========================================
# Observability
# ===========================================
structlog==24.1.0
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0

# ===========================================
# Evaluation & Metrics (RAG Evaluation Framework)
# ===========================================
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
rouge-score==0.1.2
nltk==3.8.1
bert-score==0.3.13
tabulate==0.9.0

# ===========================================
# Testing
# ===========================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
respx==0.20.2

# ===========================================
# Development & Utilities
# ===========================================
python-dotenv==1.0.0
rich==13.7.0
typer==0.9.0
